backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-jose[cryptography]
python-multipart
pytest
